GLUtilities

Documentation for GLUtilities.

GLUtilities._intervals_are_ordered

GLUtilities._pairwise_idx

GLUtilities.bin_bounds

GLUtilities.bin_center

GLUtilities.clip_interval_duration

GLUtilities.clip_ndx

GLUtilities.clip_ndx_deviance

GLUtilities.copy_length_check

GLUtilities.datevec_to_precisedatetime

GLUtilities.filter_no_collisions

GLUtilities.filtermap

GLUtilities.find_closest

GLUtilities.find_not_unique

GLUtilities.glhist

GLUtilities.glhist!

GLUtilities.interval_indices

GLUtilities.interval_intersections

GLUtilities.intervals_diff

GLUtilities.join_intervals

GLUtilities.join_intervals!

GLUtilities.moving_sum

GLUtilities.moving_sum!

GLUtilities.n_ndx

GLUtilities.ndx_offset

GLUtilities.ndx_to_t

GLUtilities.only_matches

GLUtilities.pairwise_idx

GLUtilities.pairwise_idxs

GLUtilities.parse_ranges_str

GLUtilities.skipoftype

GLUtilities.t_sup_to_ndx

GLUtilities.t_to_last_ndx

GLUtilities.t_to_ndx

GLUtilities.throttle

GLUtilities.time_interval

GLUtilities.trailing_zeros_idx

GLUtilities.typemmap

GLUtilities.window_counts

GLUtilities.window_counts

GLUtilities._intervals_are_ordered — Method

Checks that each interval is well ordered, the set of intervals is sorted, and the intervals are non-overlapping. Returns a boolean.

GLUtilities._pairwise_idx — Method

n is the number of elements in the INPUT of pairwise diff

GLUtilities.bin_bounds — Function

Find the indices to select all points in a bin

GLUtilities.bin_center — Function

Find the center index of a bin

GLUtilities.clip_interval_duration — Method

Clip an interval while trying to maintain its duration

GLUtilities.clip_ndx — Function

Clips an index to be within the valid range for an array of length l

GLUtilities.clip_ndx_deviance — Method

clip_ndx but also return deviance

GLUtilities.copy_length_check — Function

copylengthcheck returns true if dest can accept all data from source

GLUtilities.datevec_to_precisedatetime — Method

datevec_to_precisedatetime(datevec::Array{T}) where {T<:Real}

Convert matlab datevec into PreciseDateTime

GLUtilities.filter_no_collisions — Method

filter_no_collisions(as, bs, coll_rad)

Filter elements of as to keep elements that are not within coll_rad of any element in bs. Assumes both are sorted.

GLUtilities.filtermap — Method

filtermap(p::Function, f::Function, xs)

Filters the input vector, then maps the remaining values. For each element of xs which predicate function p returns true for, use mapping function f to transform the result.

GLUtilities.find_closest — Function

find_closest(a, target, ...)

Find the index of the element in a that minimizes the absolute difference from the target.

GLUtilities.find_not_unique — Method

find_not_unique(a::AbstractArray)

Returns the indices of all redundant elements in a. The time a value is seen, it is not considered redundant

GLUtilities.glhist! — Method

glhist!(cnts, xs, r)

Histogram, left inclusive. Assumes regular bin size.

GLUtilities.glhist — Method

glhist([::Type{T} = Int,] xs, r) where T

Like glhist!.

GLUtilities.interval_indices — Method

interval_indices(
    basis::Union{<:AbstractVector, AbstractRange}, start::Number, stop::Number
) -> i_b, i_e

Find the indices in basis that correspond to the interval specified by start and stop.

GLUtilities.interval_intersections — Method

Assumes each list is sorted and non-overlapping

GLUtilities.intervals_diff — Method

intervals_diff(ints_a, ints_b) -> ints_out

Find the 'set diff' of the intervals in ints_a and ints_b. Assumes both inputs are sorted.

GLUtilities.join_intervals! — Function

join_intervals!(ints::Vector{NTuple{2, <:Number}}, min_gap)

Join a list of sorted intervals, ints, if the gap between successive intervals is less than min_gap. Mutates input in-place

Assumes ints are sorted by their first index.

GLUtilities.join_intervals — Method

join_intervals(f, ints::Vector{NTuple{2, <:Number}}, min_gap)

Like join_intervals!, but does not mutate input.

GLUtilities.moving_sum! — Method

moving_sum!(out, s, nav)

Sum s in a sliding window of nav points, placing the result into out. The length of out should be max(length(s) - nav + 1, 0) if nav > 0, or length(s) otherwise.

Does not zero-pad.

GLUtilities.moving_sum — Method

moving_sum(s, nav)

Same as moving_sum!, but returns a new array.

GLUtilities.n_ndx — Method

n_ndx(start_idx::T, stop_idx::T) where {T<:Integer}

Find the number of indices between start_idx and stop_idx.

GLUtilities.ndx_offset — Method

ndx_offset(start_ndx, npt)

Finds the index that will select npt number of elements starting at start_ndx.

If npt is negative, then start_ndx is treated like the end of a range of elements, and the index required to return npt number of elements is returned.

GLUtilities.ndx_to_t — Function

ndx_to_t(i, fs, start_t)

Converts an index to its time in a regularly sampled time series.

GLUtilities.only_matches — Method

only_matches(reg::Regex, strs::AbstractArray{<:AbstractString})

Return all matches of reg in strs, skipping over non-matches.

GLUtilities.pairwise_idx — Method

n_el is the number of elements in the INPUT to pairwise diff

GLUtilities.pairwise_idxs — Method

pairwise_idxs(n::Integer) -> Vector{NTuple{2, Int}}

Returns the possible combinations of 1:n indices excluding self-pairs. The first index is always greater than the second, allowing for easy subtraction of ordered lists.

Examples

julia> pairwise_idxs(3)
3-element Vector{Tuple{Int64,Int64}}:
 (2, 1)
 (3, 1)
 (3, 2)

GLUtilities.parse_ranges_str — Method

parse_ranges_str(s::AbstractString)

Parse strings like "3", "1-5", or "1, 2-10, 12-30" into a vector of the integers in specified ranges.

GLUtilities.skipoftype — Method

skipoftype(::Type{T}, itr)

Return an iterator over the elements in itr skipping values of type T. Use collect to obtain an Array containing the non-T values in itr. Note that even if itr is a multidimensional array, the result will always be a Vector since it is not possible to remove nothings while preserving dimensions of the input.

Examples

julia> sum(GLUtilities.skipoftype(nothing, [1, nothing, 2]))
3
julia> collect(GLUtilities.skipoftype(nothing, [1, nothing, 2]))
2-element Vector{Int64}:
 1
 2
julia> collect(GLUtilities.skipoftype(nothing, [1 nothing; 2 nothing]))
2-element Vector{Int64}:
 1
 2

GLUtilities.t_sup_to_ndx — Method

t_sup_to_ndx()

Converts a time in a regularly sampled time series to the index of the last sample before the specified time.

GLUtilities.t_to_last_ndx — Function

Like ttondx, but returns sample at or before time

GLUtilities.t_to_ndx — Function

t_to_ndx()

Converts a time in a regularly sampled time series to its index. Returns the index of the first sample at or after the time specified

GLUtilities.throttle — Method

throttle(xs::AbstractVector{T}, min_gap::Number) where T<:Number ->
Vector{NTuple{2, T}}

Join points in x into ranges, if the difference between neighboring elements is less than min_gap. Assumes x is sorted.

GLUtilities.time_interval — Function

Find the time_interval of a regularly sampled time series

GLUtilities.trailing_zeros_idx — Method

trailing_zeros_idx(arr)

return last index that is not zero

GLUtilities.typemmap — Method

Make a mmaped array of type A

GLUtilities.window_counts — Method

window_counts in a certain range`

GLUtilities.window_counts — Method

window_counts(ts, window_dur)

For each event in ts, count the number of events in ts that are in the range of ts[i] and ts[i] + window_dur. Assumes ts is sorted, and that elements of ts are unique.